Fix clang warnings
* libwget/css.c (wget_css_parse_buffer): Explicitly typecast 'len' to int
* libwget/http.c (_send_callback): Change datatype of 'rc' from int to ssize_t
* libwget/http.c (_recv_callback): Likewise
* libwget/http.c (wget_http_match_no_proxy): Change param name from 'no_proxies' to
  'no_proxies_vec' to not to shadow global 'no_proxies'
* libwget/ssl_gnutls.c (check_ocsp_response): Explicit type conversion to unsigned int
* libwget/net.c (wget_tcp_listen): Move declaration of variable 'on' in #ifdef block where it's used
* src/options.c (set_long_option): Change datatypes of 'invert', 'char_insensitive'
  and 'value_present' from int to char

Reported-by: Avinash Sonawane

+++ libwget/css.c
@@ -81,7 +81,7 @@ void wget_css_parse_buffer(
 	yyscan_t scanner;
 
 	yylex_init(&scanner);
-	yy_scan_bytes(buf, len, scanner);
+	yy_scan_bytes(buf, (int) len, scanner);
 
 	while ((token = yylex(scanner)) != CSSEOF) {
 		if (token == IMPORT_SYM) {
